Cores are >= max_intensity not <=
I can see why the greater-than-or-equal sign looks wrong, because one is
asking to find pixels greater than a "maximum". However, that's okay.
You actually have pixels greater than the maximum. Maximum isn't the
maximum in the image, it is the minimum intensity of a convective core.

This needs to be reverted to >=. Not <= . The convective cores are
indeed >= max_intensity.
I found this bug after the watershed approach started grouping the whole
image into a single label.